+++ ui/app/src/components/FormsEngine/controls/NodeSelector.tsx
@@ -33,7 +33,7 @@ import { FormattedMessage } from 'react-intl';
 import LinkOffRoundedIcon from '@mui/icons-material/LinkOffRounded';
 import Tooltip from '@mui/material/Tooltip';
 import useContentTypes from '../../../hooks/useContentTypes';
-import {
+import React, {
 	lazy,
 	MouseEvent as ReactMouseEvent,
 	ReactNode,
@@ -78,7 +78,7 @@ import useItemsByPath from '../../../hooks/useItemsByPath';
 import ItemDisplay from '../../ItemDisplay';
 import useActiveUser from '../../../hooks/useActiveUser';
 import { getFileExtension, processPathMacros } from '../../../utils/path';
-import { ensureSingleSlash } from '../../../utils/string';
+import { ensureSingleSlash, isEmpty } from '../../../utils/string';
 import { popDialog, pushDialog } from '../../../state/actions/dialogStack';
 import FieldBox from '../components/FieldBox';
 import { isTouchDevice, KeyDownEvent, sortableListKeyDownHandler } from '../lib/sortableListUtil';
@@ -106,6 +106,10 @@ import {
 } from '../../PathNavigator/utils';
 import useSelection from '../../../hooks/useSelection';
 import { getValidationValue, isFieldReadOnly, showAlert } from '../lib/formUtils';
+import TypeList from '../../ContentTypeManagement/components/TypeList';
+import { SearchBar } from '../../SearchBar';
+import useDebouncedInput from '../../../hooks/useDebouncedInput';
+import { filterTypesByKeywordsAndObjectType } from '../../../utils/contentType';
 
 const SortableList = lazy(() => import('../components/SortableList'));
 const TouchSortableList = lazy(() => import('../components/TouchSortableList'));
@@ -508,7 +512,17 @@ function NodeSelector(props: NodeSelectorProps) {
 				}}
 				children={menuOptions}
 			/>
-			<Dialog open={pickerDialogOpen} onClose={handleCloseDataSourcePickerDialog} fullWidth maxWidth="sm">
+			<Dialog
+				open={pickerDialogOpen}
+				onClose={handleCloseDataSourcePickerDialog}
+				fullWidth
+				maxWidth={pickerType !== 'create' ? 'sm' : 'md'}
+				slotProps={{
+					paper: {
+						sx: [pickerType === 'create' && { minHeight: '60vh' }]
+					}
+				}}
+			>
 				<DialogHeader
 					title={<FormattedMessage defaultMessage="Choose how to proceed" />}
 					onCloseButtonClick={handleCloseDataSourcePickerDialog}
@@ -764,8 +778,8 @@ function CreateDataSourcePicker(props: {
 	});
 	const refs = useUpdateRefs({ value, onChange });
 	const [allowedStrategies, setAllowedStrategies] = useState({ embedded: false, shared: false });
-	const handleTypeChange = (event: SyntheticEvent) => {
-		const newValue = { ...value, contentTypeId: (event.target as HTMLInputElement).value };
+	const handleTypeChange = (event: SyntheticEvent, contentType: ContentType) => {
+		const newValue = { ...value, contentTypeId: contentType.id };
 		setValue(newValue);
 		onChange?.(event, newValue);
 	};
@@ -831,47 +845,36 @@ function CreateDataSourcePicker(props: {
 	useEffect(() => {
 		refs.current.onChange?.(null, value);
 	}, [refs, value]);
+
+	const [filteredTypes, setFilteredTypes] = useState<ContentType[] | undefined>(undefined);
+	const [keywords, setKeywords] = useState<string>('');
+	const onKeyword$ = useDebouncedInput((keywords) => {
+		if (!allowedTypes) return;
+		const types = filterTypesByKeywordsAndObjectType(
+			allowedTypes.map((typeId) => contentTypesLookup[typeId]).filter(Boolean),
+			keywords,
+			'all'
+		);
+		setFilteredTypes(types);
+	});
+
+	useEffect(() => {
+		if (allowedTypes?.length) {
+			setFilteredTypes((current) =>
+				current === undefined ? allowedTypes.map((typeId) => contentTypesLookup[typeId]).filter(Boolean) : current
+			);
+		}
+	}, [allowedTypes, contentTypesLookup]);
+
+	const handleKeywordsChange = (value: string) => {
+		setKeywords(value);
+		onKeyword$.next(value);
+	};
+
 	return (
-		<Grid container spacing={2} justifyContent="center">
-			<Grid sx={{ display: 'flex', flexDirection: 'column' }}>
-				<FormControl>
-					<FormLabel id="contentTypeLabel" sx={{ minHeight: 28, display: 'flex', alignItems: 'center' }}>
-						<FormattedMessage defaultMessage="Content Type" />
-					</FormLabel>
-					<RadioGroup aria-labelledby="contentTypeLabel" name="contentType" value={value.contentTypeId}>
-						{allowedTypes?.map((contentTypeId) => (
-							<FormControlLabel
-								key={contentTypeId}
-								value={contentTypeId}
-								control={<Radio />}
-								label={contentTypesLookup[contentTypeId].name}
-								onChange={handleTypeChange}
-							/>
-						))}
-					</RadioGroup>
-				</FormControl>
-				{props.allowedCreateTypes[value.contentTypeId]?.createPaths?.length > 1 && (
-					<FormControl sx={{ mt: 1 }}>
-						<FormLabel>
-							<FormattedMessage defaultMessage="Creation Path" />
-						</FormLabel>
-						<RadioGroup aria-labelledby="creationPathLabel" name="creationPath" value={value.path} sx={{}}>
-							{props.allowedCreateTypes[value.contentTypeId].createPaths.map((path) => (
-								<FormControlLabel
-									key={path}
-									value={path}
-									control={<Radio />}
-									onChange={handlePathChange}
-									label={<Typography noWrap maxWidth="100%" component="div" title={path} children={path} />}
-									disableTypography
-								/>
-							))}
-						</RadioGroup>
-					</FormControl>
-				)}
-			</Grid>
+		<Grid container spacing={2} display="flex" flexDirection="column">
 			<Grid>
-				<FormControl sx={{ mb: 1, shrink: 0 }}>
+				<FormControl sx={{ mb: 1, flexShrink: 0 }} fullWidth>
 					<Box alignItems="center" display="flex">
 						<FormLabel id="creationStrategyLabel">
 							<FormattedMessage defaultMessage="Creation Strategy" />
@@ -880,7 +883,7 @@ function CreateDataSourcePicker(props: {
 							<HelpOutline fontSize="inherit" />
 						</IconButton>
 					</Box>
-					<RadioGroup aria-labelledby="creationStrategyLabel" name="creationStrategy" value={value.strategy}>
+					<RadioGroup aria-labelledby="creationStrategyLabel" name="creationStrategy" value={value.strategy} row>
 						<FormControlLabel
 							value="embedded"
 							disabled={!allowedStrategies.embedded}
@@ -894,6 +897,59 @@ function CreateDataSourcePicker(props: {
 							label={<FormattedMessage defaultMessage="Shared" />}
 						/>
 					</RadioGroup>
+				</FormControl>
+			</Grid>
+			{value.strategy === 'shared' && allowedCreateTypes[value.contentTypeId]?.createPaths?.length > 1 && (
+				<Grid sx={{ display: 'flex', flexDirection: 'column' }}>
+					<FormControl sx={{ mt: 1 }} fullWidth>
+						<FormLabel id="creationPathLabel">
+							<FormattedMessage defaultMessage="Creation Path" />
+						</FormLabel>
+						<RadioGroup aria-labelledby="creationPathLabel" name="creationPath" value={value.path} sx={{}}>
+							{allowedCreateTypes[value.contentTypeId].createPaths.map((path) => (
+								<FormControlLabel
+									key={path}
+									value={path}
+									control={<Radio />}
+									onChange={handlePathChange}
+									label={<Typography noWrap maxWidth="100%" component="div" title={path} children={path} />}
+									disableTypography
+								/>
+							))}
+						</RadioGroup>
+					</FormControl>
+				</Grid>
+			)}
+
+			<Grid width="100%">
+				<FormControl fullWidth>
+					<FormLabel id="contentTypeLabel" sx={{ minHeight: 28, display: 'flex', alignItems: 'center' }}>
+						<FormattedMessage defaultMessage="Content Type" />
+					</FormLabel>
+					<SearchBar
+						keyword={keywords}
+						onChange={(value) => handleKeywordsChange(value)}
+						showActionButton={!isEmpty(keywords)}
+						sxs={{
+							root: {
+								background: 'none !important',
+								border: 'none !important',
+								borderRadius: 0,
+								boxShadow: 'none',
+								flexGrow: 1,
+								py: 1
+							},
+							inputInput: { padding: '8px 5px' }
+						}}
+					/>
+					<TypeList
+						contentTypes={filteredTypes}
+						compact={true}
+						onCardClick={handleTypeChange}
+						selectedTypeId={value.contentTypeId}
+						disableSelected={false}
+						skeleton={filteredTypes === undefined}
+					/>
 				</FormControl>
 			</Grid>
 		</Grid>