+++ ui/app/src/components/PreviewConcierge/PreviewConcierge.tsx
@@ -1109,7 +1109,8 @@ export function PreviewConcierge(props: PropsWithChildren<{}>) {
 							props: {
 								item: payload.item,
 								onClosed: () => dispatch(requestWorkflowCancellationDialogOnResult({ type: 'close' })),
-								onContinue: () => dispatch(requestWorkflowCancellationDialogOnResult({ type: 'continue' }))
+								onContinue: (cancelPackagesComment) =>
+									dispatch(requestWorkflowCancellationDialogOnResult({ type: 'continue', cancelPackagesComment }))
 							}
 						})
 					);