+++ ui/app/src/components/UserManagement/UserManagement.tsx
@@ -107,10 +107,6 @@ export function UserManagement(props: UserManagementProps) {
 		setLimit(e.target.value);
 	};
 
-	const onSearchButtonClick = () => {
-		searchInpuRef.current?.focus();
-	};
-
 	const onSearch = useCallback(
 		(keyword) => {
 			fetchUsers(keyword, 0);
@@ -158,7 +154,6 @@ export function UserManagement(props: UserManagementProps) {
 						}}
 						keyword={keyword}
 						onChange={handleSearchKeyword}
-						onDecoratorButtonClick={onSearchButtonClick}
 						showActionButton={Boolean(keyword)}
 					/>
 				}