+++ ui/app/src/components/ViewPackagesDialog/ViewPackagesDialog.tsx
@@ -24,10 +24,11 @@ import { ContentItem } from '../../models';
 
 export interface ViewPackagesDialogBaseProps {
 	item: ContentItem;
+	cancelPackagesInitialComment?: string;
 }
 
 export interface ViewPackagesDialogProps extends ViewPackagesDialogBaseProps, EnhancedDialogProps {
-	onContinue?(): void;
+	onContinue?(cancelPackagesMessage: string): void;
 }
 
 export interface ViewPackagesDialogStateProps extends ViewPackagesDialogBaseProps, EnhancedDialogState {
@@ -37,7 +38,7 @@ export interface ViewPackagesDialogStateProps extends ViewPackagesDialogBaseProp
 }
 
 export function ViewPackagesDialog(props: ViewPackagesDialogProps) {
-	const { item, onContinue, ...enhancedDialogProps } = props;
+	const { item, cancelPackagesInitialComment, onContinue, ...enhancedDialogProps } = props;
 	return (
 		<EnhancedDialog
 			fullWidth
@@ -48,7 +49,11 @@ export function ViewPackagesDialog(props: ViewPackagesDialogProps) {
 			}
 			{...enhancedDialogProps}
 		>
-			<ViewPackagesDialogContainer item={item} onContinue={onContinue} />
+			<ViewPackagesDialogContainer
+				item={item}
+				cancelPackagesInitialComment={cancelPackagesInitialComment}
+				onContinue={onContinue}
+			/>
 		</EnhancedDialog>
 	);
 }