+++ ui/app/src/utils/itemActions.ts
@@ -700,7 +700,8 @@ export const itemActionDispatcher = ({
 									dispatch(pushErrorDialog({ props: { error: response } }));
 								}
 							});
-						}
+						},
+						formatMessage
 					});
 				}
 				break;